fixed everething but the delete popup
