+++ src/blocks/element/__likes-number/element__likes_number.css
@@ -1,5 +1,4 @@
 .element__likes-number {
-  visibility: visible;
   box-sizing: border-box;
   font-size: 13px;
   font-weight: 400;
