+++ src/components/PopupWithDelete.js
@@ -1,13 +1,33 @@
 import Popup from "./Popup.js";
 
 export default class PopupWithDelete extends Popup  {
-  constructor(popupSelector, callback) {
+  constructor(popupSelector, {deleteCallback}) {
     super(popupSelector);
-    this._callback = callback;
-  }
-setEventListeners() {
+    this._deleteCallback = deleteCallback;
+    this._button = this._popupSelector.querySelector('#confirmDelete');
+    this._functionForDelete = this._functionForDelete.bind(this);
+    }
+
+    openPopup(evt) {
+      super.openPopup();
+      this._cardToDelete = evt.target.closest('.element');
+      this.setEventListeners();
+    }
+
+
+  setEventListeners() {
   super.setEventListeners();
-  this._cardForDelete = evt.target.closest('.element');
-  this._popupSelector.querySelector('#confirmDelete').addEventListener('click', () => { this._callback(this._cardToDelete) } );
+  this._button.addEventListener('click', this._functionForDelete);
 }
+
+_functionForDelete() {
+  this._deleteCallback(this._cardToDelete);
 }
+
+removeEventListeners() {
+  this._confirmButton.removeEventListener('click', this._functionForDelete);
+}
+
+}
+
+
